requirements.txt
streamlit
pyvis
networkx
google-genai